requirements.txt
# Core Dependencies
Flask==2.3.3
opencv-python==4.8.0.76
numpy==1.24.3
tensorflow==2.13.0
cvzone==1.5.6
python-dotenv==1.0.0

# Web/Server
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# TensorFlow Dependencies
h5py==3.9.0
absl-py==1.4.0
astunparse==1.6.3
flatbuffers==23.5.26
gast==0.4.0
google-pasta==0.2.0
grpcio==1.57.0
libclang==16.0.0
opt-einsum==3.3.0
protobuf==4.23.4
six==1.16.0
tensorboard==2.13.0
tensorflow-estimator==2.13.0
termcolor==2.3.0
typing-extensions==4.7.1
wrapt==1.14.1
tensorflow-io-gcs-filesystem==0.31.0